+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 
 ## Improvements
 - Added an easy way to open replays folder by clicking on the replay saved message, or executing `/guesstheutils replay open`
+- Changed theme length display to individual word lengths